+++ android/src/oldarch/java/com/reactnativestripesdk/NativeOnrampSdkModuleSpec.java
@@ -45,6 +45,10 @@ public abstract class NativeOnrampSdkModuleSpec extends ReactContextBaseJavaModu
 
   @ReactMethod
   @DoNotStrip
+  public abstract void isSamsungPaySupported(Promise promise);
+
+  @ReactMethod
+  @DoNotStrip
   public abstract void hasLinkAccount(String email, Promise promise);
 
   @ReactMethod